Corrected problems from review
