+++ packages/xrpl-did-resolver/src/utils/jsonUtils.ts
@@ -1,0 +1,14 @@
+import { Errors } from './errors'
+
+export async function parseAndValidateJson(res: Response): Promise<any> {
+  let doc: unknown
+  try {
+    doc = await res.json()
+  } catch {
+    throw new Error(Errors.invalidJson)
+  }
+  if (typeof doc !== 'object' || doc === null || Array.isArray(doc)) {
+    throw new Error(Errors.invalidJson)
+  }
+  return doc
+}
